+++ docs/include/samplecode/SpotifyClient/IsChapterEpisode.py
@@ -1,0 +1,26 @@
+from spotifywebapipython import *
+from spotifywebapipython.models import *
+
+try:
+
+    CLIENT_ID:str = 'your_client_id'
+    CLIENT_SECRET:str = 'your_client_secret'
+
+    # create new spotify client instance.
+    spotify:SpotifyClient = SpotifyClient()
+
+    # generate a spotify client credentials access token (no scope, public data use only).
+    spotify.SetAuthTokenClientCredentials(CLIENT_ID, CLIENT_SECRET)
+    print('\nAuth Token:\n Type="%s"\n Scope="%s"' % (spotify.AuthToken.AuthorizationType, str(spotify.AuthToken.Scope)))
+    print('\nUser:\n DisplayName="%s"\n EMail="%s"' % (spotify.UserProfile.DisplayName, spotify.UserProfile.EMail))
+
+    # is episode id an audiobook chapter?
+    episodeId:str = '1mbA4ji7bRh83EqxWFni3L'
+    print('\nChecking if episode ID is an audiobook chapter: \n- %s \n' % episodeId)
+    result:bool = spotify.IsAudiobookEpisode(episodeId)
+
+    print('IsAudiobookEpisode = %s' % str(result))
+
+except Exception as ex:
+
+    print("** Exception: %s" % str(ex))