+++ CHANGELOG.md
@@ -6,6 +6,10 @@ Change are listed in reverse chronological order (newest to oldest).
 
 <span class="changelog">
 
+###### [ 1.0.98 ] - 2024/09/24
+
+  * Added `SpotifyClient.IsAudiobookEpisode` method that returns true if the specified episode id is an audiobook chapter; otherwise, false.  This allows you to programatically determine the difference between a podcast / show episode, and an audiobook episode.
+
 ###### [ 1.0.97 ] - 2024/09/20
 
   * Updated `ZeroconfGetInfoAlias` to fix a bug related to Spotify Connect devices that define aliases (introduced with v1.0.93).  This was causing `TypeError: unsupported operand type(s) for &: 'str' and 'tuple'` exceptions if a device had aliases defined.